Adds paywallClose paywall event receiver capability
