+++ Sources/SuperwallKit/Paywall/Capabilities/Array+Capability.swift
