+++ Sources/SuperwallKit/Paywall/View Controller/PaywallViewController.swift
@@ -250,6 +250,7 @@ public class PaywallViewController: UIViewController, LoadingDelegate {
       paywallInfo: info,
       surveyPresentationResult: surveyPresentationResult
     )
+    await webView.messageHandler.handle(.paywallClose)
     await Superwall.shared.track(trackedEvent)
     await triggerSessionManager.trackPaywallClose()
   }
