+++ Sources/SuperwallKit/Paywall/View Controller/Web View/Message Handling/PaywallMessage.swift
@@ -47,6 +47,7 @@ enum PaywallMessage: Decodable, Equatable {
   case custom(data: String)
 
   // All cases below here are sent from device to paywall
+  case paywallClose
   case paywallOpen
 
   case restoreStart
